split mock handler
